+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Consolidate `CacheVersions` and bump to refresh `CacheKey` usage. ([#1041](https://github.com/getsentry/symbolicator/pull/1041), [#1042](https://github.com/getsentry/symbolicator/pull/1042))
 - Automatically block downloads from unreliable hosts. ([#1039](https://github.com/getsentry/symbolicator/pull/1039))
 - Fully migrate `CacheKey` usage and remove legacy markers. ([#1043](https://github.com/getsentry/symbolicator/pull/1043))
+- Add support for in-memory caching. ([#1028](https://github.com/getsentry/symbolicator/pull/1028))
 
 ### Fixes